refac: suggestions view file
